+++ CHANGELOG.md
@@ -17,6 +17,10 @@ and this project adheres to
 
 - ♻️(tilt) use helm dev-backend chart
 
+### Fixed
+
+- 🐛(front) fix the content when opening the right panel #328
+
 ## [v0.4.0] - 2025-09-02
 
 ### Added
@@ -46,7 +50,6 @@ and this project adheres to
 - 🐛(backend) allow item partial update without modifying title #316
 - 🌐(front) fix share item wording #315
 - 🐛(front) fix left tree glitch #323
-
 
 ## [v0.2.0] - 2025-08-18